+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/intel/requests/UserBreachedRequest.java
@@ -30,6 +30,10 @@ public class UserBreachedRequest extends IntelCommonRequest {
 	@JsonProperty("end")
 	String end;
 
+	@JsonInclude(Include.NON_NULL)
+	@JsonProperty("cursor")
+	String cursor;
+
 	protected UserBreachedRequest(Builder builder) {
 		super(builder);
 		this.email = builder.email;
@@ -38,6 +42,7 @@ public class UserBreachedRequest extends IntelCommonRequest {
 		this.phoneNumber = builder.phoneNumber;
 		this.start = builder.start;
 		this.end = builder.end;
+		this.cursor = builder.cursor;
 	}
 
 	public static class Builder extends CommonBuilder<Builder> {
@@ -48,6 +53,7 @@ public class UserBreachedRequest extends IntelCommonRequest {
 		String phoneNumber;
 		String start;
 		String end;
+		String cursor;
 
 		public Builder() {}
 
@@ -82,6 +88,11 @@ public class UserBreachedRequest extends IntelCommonRequest {
 
 		public Builder end(String end) {
 			this.end = end;
+			return this;
+		}
+
+		public Builder cursor(String cursor) {
+			this.cursor = cursor;
 			return this;
 		}
 	}